pt_resnet50_imagenet_224_224_8.2G_3.0/requirement_upgraded.txt
contourpy==1.3.3
cycler==0.12.1
dominate==2.9.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.9.0
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markupsafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
netron==8.8.2
networkx==3.5
numpy==2.3.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pip==25.2
protobuf==6.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchsummary==1.5.1
torchvision==0.23.0
tqdm==4.67.1
typing-extensions==4.15.0
wheel==0.45.1
